Reorganize commands for cli
